Change FixedTimeWorkout splitSize to tenths
It was measured in seconds, which turned out to be an order
of magnitude too much. We're going to have to update the dataformat.md
eventually to eliminate some of this confusion; the original document
makes no distinction between the two, and it's led to a whole bunch of
stupid bugs.