Start adding JSON encoding to DataTypes
This is going to be a long, annoying process, one that is probably
going to involve some annoyance due to the fact that calculations
will have to be done across disparate objects to get everything
that the Concept2 API needs. I'm going to need to roll some more
functions for merging objects, combining data, and so on.