Convert datestamps to proper format
The timestamps start in UTC time. We need to convert them to the local
time. We actually need to parse it from localtime as well, but
that is significantly harder - it would mean that all of the parseHeader
functions would need to include a TimeZone. We have time, of course.